requirements.txt
asgiref==3.2.7
Django==3.0.6
future==0.18.2
numpy==1.18.4
opencv-python==4.0.0.21
Pillow==7.1.2
pyclipper==1.1.0.post3
pytz==2020.1
Shapely==1.7.0
sqlparse==0.3.1
torch==1.5.0+cu101
torchvision==0.6.0+cu101
